Create README.md
This object detection is a classical problem in computer vision where you work to reognize what and where specifically what objects inside a given image and also where they are in the image the problem of object detection is more complex than classification which can recognize objects but doesn't indicate where the object is located in the image. In addition classification does not work images containing more than one object Earlier the algorithms which were implemented in two stages are slow because they select the region of interest in an image secondly they classify using convolution neural network this solution can be slow because we have to run prediction on each and every selected region so imagine if you want to extract 5000 regions for each image based on selective search how much it will take time. Suppose we have N images then the number lof CNN features will be N*5000 it increases the time complexity and also cannot be applied in real time due to low fps thats why we are migrating towards single pipeline architecture like yolo the neural network of yolo looks the entire image at single instance hence its name.